Finished clock
Followed the video demonstration from the challenge, did it in front with
good results. The end of the assignments was ver interesting, since my main
focus was on stopping the clock hands from jumping around every time they
needed to finish the full rotation.

Tried a few solution, first one being the MutationObserver Class of vanilla
JS, which proved to be a bit more complex then needed but it was a good
exercises.

Final solution was made with the special corrective function which toggles
the transition property of the hand in question.