Interactive flex panel gallery
An interesting way to create a nice effect on the with vanilla tools,
Wes had an interesting approach to the solution and I think it was more
detailed, while my solution was faster and simpler t the first glance. It
 would be interesting to learn what is the difference.

But an interesting piece of information was given, regarding function
calling inside the event listener. If as a second parameter there is a
standard function call (example()), it is run on page load, while what we
want is to reference the function by just using it's name.

Referencing the function in such a way only activates the function when
that piece of code is activated.